新增构建配置文件

+++ sapi/make.php
@@ -3,40 +3,60 @@
  * @var $this SwooleCli\Preprocessor
  */
 ?>
+set -uex
+PKG_CONFIG_PATH='/usr/lib/pkgconfig'
+test -d /usr/lib64/pkgconfig && PKG_CONFIG_PATH="/usr/lib64/pkgconfig:$PKG_CONFIG_PATH" ;
+test -d /usr/local/lib/pkgconfig && PKG_CONFIG_PATH="/usr/local/lib/pkgconfig:$PKG_CONFIG_PATH" ;
+test -d /usr/local/lib64/pkgconfig && PKG_CONFIG_PATH="/usr/local/lib64/pkgconfig:$PKG_CONFIG_PATH" ;
+
 SRC=<?= $this->phpSrcDir . PHP_EOL ?>
 ROOT=$(pwd)
 export CC=clang
 export CXX=clang++
 export LD=ld.lld
 export PKG_CONFIG_PATH=<?= implode(':', $this->pkgConfigPaths) . PHP_EOL ?>
+export ORIGIN_PKG_CONFIG_PATH=$PKG_CONFIG_PATH
 OPTIONS="--disable-all \
 <?php foreach ($this->extensionList as $item) : ?>
-<?=$item->options?> \
+    <?=$item->options?> \
 <?php endforeach; ?>
 <?=$this->extraOptions?>
 "
 
 <?php foreach ($this->libraryList as $item) : ?>
+
 make_<?=$item->name?>() {
-    cd <?=$this->workDir?>/thirdparty
-    echo "build <?=$item->name?>"
-    mkdir -p <?=$this->workDir?>/thirdparty/<?=$item->name?> && \
-    tar --strip-components=1 -C <?=$this->workDir?>/thirdparty/<?=$item->name?> -xf <?=$this->workDir?>/pool/lib/<?=$item->file?>  && \
-    cd <?=$item->name?> && \
-    echo  "<?=$item->configure?>"
+    cd <?=$this->workDir?>/thirdparty ;
+    echo "build <?=$item->name?>" ;
+    <?php if ($item->beforeConfigureCleanPackageFlag == true ): ?>
+    test -d <?=$this->workDir?>/thirdparty/<?=$item->name?> && rm -rf <?=$this->workDir?>/thirdparty/<?=$item->name?> ;
+    <?php endif;?>
+    mkdir -p <?=$this->workDir?>/thirdparty/<?=$item->name?> ;
+    tar --strip-components=1 -C <?=$this->workDir?>/thirdparty/<?=$item->name?> -xf <?=$this->workDir?>/pool/lib/<?=$item->file?> ;
+    cd <?=$item->name?> ;
+    <?php if (!empty($item->beforeConfigureScript)):?>
+        <?= $item->beforeConfigureScript ?>
+    <?php endif ;?>
+    :;
+    echo  "<?=$item->configure?>" ;
     <?php if (!empty($item->configure)): ?>
-    <?=$item->configure?> && \
+    <?= $item->configure ?> && \
     <?php endif; ?>
+    :;
     make -j <?=$this->maxJob?>  <?=$item->makeOptions?> && \
     <?php if ($item->beforeInstallScript): ?>
-    <?=$item->beforeInstallScript?> && \
+    <?=$item->beforeInstallScript?>
     <?php endif; ?>
-    make install <?=$item->makeInstallOptions?> && \
+    :;
+    make <?=$item->makeInstallDefaultOptions?> <?= $item->makeInstallOptions ?> && \
+    :;
     <?php if ($item->afterInstallScript): ?>
-    <?=$item->afterInstallScript?> && \
+    <?= $item->afterInstallScript?>
     <?php endif; ?>
+    :;
     cd -
 }
+
 
 clean_<?=$item->name?>() {
     cd <?=$this->workDir?>/thirdparty
@@ -45,6 +65,7 @@ clean_<?=$item->name?>() {
     cd -
 }
 <?php echo str_repeat(PHP_EOL, 1);?>
+
 <?php endforeach; ?>
 
 make_all_library() {
@@ -54,27 +75,68 @@ make_all_library() {
 }
 
 config_php() {
-    rm ./configure
-    ./buildconf --force
+<?php if ( $this->disableZendOpcacheFlag == true ) : ?>
+    test -f main/main.c.save ||  cp -f main/main.c main/main.c.save ;
+    sed -i 's/extern zend_extension zend_extension_entry;//g' main/main.c ;
+    sed -i 's/zend_register_extension(&zend_extension_entry, NULL);//g' main/main.c ;
+<?php else : ?>
+    test -f main/main.c.save &&  cp -f main/main.c.save main/main.c ;
+<?php endif; ?>
+     test -f ./configure && rm ./configure ;
+
+     export FREETYPE2_CFLAGS=$(pkg-config --cflags freetype2) ;
+     export FREETYPE2_LIBS=$(pkg-config --libs freetype2) ;
+
+     export LIBSODIUM_CFLAGS=$(pkg-config --cflags libsodium) ;
+     export LIBSODIUM_LIBS=$(pkg-config --libs libsodium) ;
+
+     export XSL_CFLAGS=$(pkg-config --cflags libxslt) ;
+     export XSL_LIBS=$(pkg-config --libs libxslt) ;
+
+     export ONIG_CFLAGS=$(pkg-config --cflags oniguruma) ;
+     export ONIG_LIBS=$(pkg-config --libs oniguruma) ;
+
+     export LIBZIP_CFLAGS=$(pkg-config --cflags libzip) ;
+     export LIBZIP_LIBS=$(pkg-config --libs libzip) ;
+
+:<<'EOF'
+
+     export PKG_CONFIG_PATH="/usr/icu/lib/pkgconfig:$PKG_CONFIG_PATH"
+     export ICU_CFLAGS=$(pkg-config --cflags  icu-uc icu-io icu-i18n)  ;
+     export ICU_LIBS=$(pkg-config  --libs icu-uc icu-io icu-i18n)  ;
+
+     export PKG_CONFIG_PATH="/usr/ncurses/lib/pkgconfig:/usr/readline/lib/pkgconfig:$PKG_CONFIG_PATH"
+
+     export NCURSES_CFLAGS=$(pkg-config --cflags ncursesw ncurses);
+     export NCURSES_LIBS=$(pkg-config  --libs ncursesw ncurses);
+
+     export READLINE_CFLAGS=$(pkg-config --cflags  readline)  ;
+     export READLINE_LIBS=$(pkg-config  --libs readline)  ;
+
+EOF
+
+
+    ./buildconf --force ;
+
 <?php if ($this->osType !== 'macos') : ?>
     mv main/php_config.h.in /tmp/cnt
     echo -ne '#ifndef __PHP_CONFIG_H\n#define __PHP_CONFIG_H\n' > main/php_config.h.in
     cat /tmp/cnt >> main/php_config.h.in
     echo -ne '\n#endif\n' >> main/php_config.h.in
 <?php endif; ?>
-    echo $OPTIONS
-    echo $PKG_CONFIG_PATH
-    ./configure $OPTIONS
+    echo $OPTIONS ;
+    echo $PKG_CONFIG_PATH ;
+    ./configure $OPTIONS ;
 }
 
 make_php() {
     make EXTRA_CFLAGS='-fno-ident -Xcompiler -march=nehalem -Xcompiler -mtune=haswell -Os' \
-    EXTRA_LDFLAGS_PROGRAM='-all-static -fno-ident <?=$this->extraLdflags?> <?php foreach ($this->libraryList as $item) {
-        if (!empty($item->ldflags)) {
-            echo $item->ldflags;
-            echo ' ';
-        }
-    } ?>'  -j <?=$this->maxJob?> && echo ""
+         EXTRA_LDFLAGS_PROGRAM='-all-static -fno-ident <?=$this->extraLdflags?> <?php foreach ($this->libraryList as $item) {
+    if (!empty($item->ldflags)) {
+        echo $item->ldflags;
+        echo ' ';
+    }
+} ?>'  -j <?=$this->maxJob?> && echo ""
 }
 
 help() {
@@ -88,9 +150,9 @@ help() {
 }
 
 if [ "$1" = "docker-build" ] ;then
-  sudo docker build -t phpswoole/swoole_cli_os:<?= $this->dockerVersion ?> .
+    sudo docker build -t phpswoole/swoole_cli_os:<?= $this->dockerVersion ?> .
 elif [ "$1" = "docker-bash" ] ;then
-    sudo docker run -it -v $ROOT:<?=$this->workDir?> phpswoole/swoole_cli_os:<?= $this->dockerVersion ?> /bin/bash
+    sudo docker run -it -v $ROOT:<?=$this->workDir?> --workdir <?=$this->workDir?> phpswoole/swoole_cli_os:<?= $this->dockerVersion ?> /bin/bash
     exit 0
 elif [ "$1" = "all-library" ] ;then
     make_all_library
@@ -117,7 +179,7 @@ elif [ "$1" = "clean-all-library" ] ;then
     clean_<?=$item->name?> && echo "[SUCCESS] make clean [<?=$item->name?>]"
 <?php endforeach; ?>
 elif [ "$1" = "diff-configure" ] ;then
-  meld $SRC/configure.ac ./configure.ac
+    meld $SRC/configure.ac ./configure.ac
 elif [ "$1" = "pkg-check" ] ;then
 <?php foreach ($this->libraryList as $item) : ?>
     echo "[<?= $item->name ?>]"
@@ -158,6 +220,7 @@ elif [ "$1" = "sync" ] ;then
   cp -r $SRC/ext/pcre/ ./ext
   cp -r $SRC/ext/pdo/ ./ext
   cp -r $SRC/ext/pdo_mysql/ ./ext
+  cp -r $SRC/ext/pdo_pgsql/ ./ext
   cp -r $SRC/ext/pdo_sqlite/ ./ext
   cp -r $SRC/ext/phar/ ./ext
   cp -r $SRC/ext/posix/ ./ext
@@ -169,6 +232,7 @@ elif [ "$1" = "sync" ] ;then
   cp -r $SRC/ext/sockets/ ./ext
   cp -r $SRC/ext/sodium/ ./ext
   cp -r $SRC/ext/spl/ ./ext
+  cp -r $SRC/ext/skeleton/ ./ext
   cp -r $SRC/ext/sqlite3/ ./ext
   cp -r $SRC/ext/standard/ ./ext
   cp -r $SRC/ext/sysvshm/ ./ext
@@ -193,4 +257,5 @@ elif [ "$1" = "sync" ] ;then
 else
     help
 fi
+
